refactor: 닉네임 중복여부 usecase 스코프 변경
sign up feature에서 user feature로 변경
