+++ lib/features/sign_up/data/remote/sign_up_remote_data_source.dart
@@ -1,3 +1,1 @@
-abstract interface class SignUpRemoteDataSource {
-  Future<bool> isExistNickname(String nickname);
-}
+abstract interface class SignUpRemoteDataSource {}
