+++ lib/features/sign_up/repositories/sign_up_repository.dart
@@ -1,3 +1,1 @@
-abstract interface class SignUpRepository {
-  Future<bool> isExistNickname(String nickname);
-}
+abstract interface class SignUpRepository {}
