+++ lib/features/sign_up/repositories/sign_up_repository_impl.dart
@@ -7,9 +7,4 @@ final class SignUpRepositoryImpl implements SignUpRepository {
   );
 
   final SignUpRemoteDataSource _signUpRemoteDataSource;
-
-  @override
-  Future<bool> isExistNickname(String nickname) {
-    return _signUpRemoteDataSource.isExistNickname(nickname);
-  }
 }
