+++ lib/features/sign_up/sign_up.dart
@@ -1,3 +1,3 @@
+export '../user/usecases/is_exist_nickname_use_case.dart';
 export 'data/remote/sign_up_remote_data_source.dart';
 export 'repositories/sign_up_repository.dart';
-export 'usecases/is_exist_nickname_use_case.dart';
