+++ lib/features/user/data/remote/user_remote_data_source.dart
@@ -3,4 +3,5 @@ import 'package:techtalk/features/user/models/user_data_model.dart';
 abstract interface class UserRemoteDataSource {
   Future<void> createUserData(UserDataModel data);
   Future<UserDataModel?> getUserData(String uid);
+  Future<bool> isExistNickname(String nickname);
 }
