+++ lib/features/user/usecases/is_exist_nickname_use_case.dart
@@ -1,0 +1,13 @@
+import 'package:techtalk/features/user/user.dart';
+
+final class IsExistNicknameUseCase {
+  const IsExistNicknameUseCase(
+    this._userRepository,
+  );
+
+  final UserRepository _userRepository;
+
+  Future<bool> call(String nickname) async {
+    return _userRepository.isExistNickname(nickname);
+  }
+}
