+++ lib/features/user/user.dart
@@ -2,3 +2,4 @@ export 'data/remote/user_remote_data_source.dart';
 export 'repositories/user_repository.dart';
 export 'usecases/create_user_data_use_case.dart';
 export 'usecases/get_user_data_use_case.dart';
+export 'usecases/is_exist_nickname_use_case.dart';
